Code coverage testing completion
Frequent disconnection of tunnels was solved by using a local identifier
for all tests.

New current issues with browsers (mostly BrowserStack issues) were seen,
and whatever is not consistently working (not the tunnel break issue
anymore) has been removed from testing.

As of now only Android 4.3 and 4.4 browsers are showing some issue that
points to something that has to do with our test/source code. They need
to be debugged and for now have been removed from build.

Code coverage shows up in codecov/coveralls.

README updated with latest browser support status with browsers that are
working fine. As of now none of the tests are failing anywhere, which
means all APIs are working on whichever browsers are working.

[ci skip]