Bug fixes
Closes #7
Closes #8
Closes #9

MDN can be incorrect at times. The 'value' member does not exist in
DOMTokenList (classList member of element) for several old and pretty
recent browsers, contrary to how MDN says the browser support for this
member is.

Since 'value' did not exist, several test cases across several browsers
failed for getClass and its use in tests for addClass and removeClass.
Fix was to provide the required value using Array.prototype.join in case
the 'value' member does not exist.